# src/modules/order/dto/create-order.dto.ts
import { ApiProperty } from "@nestjs/swagger";
import { IsEmail, IsNotEmpty, IsNumber, IsPhoneNumber, IsString } from "class-validator";
import { OrderItemEntity } from "../../../entities/order-item.entity";

export class CreateOrderDto {

    @ApiProperty()
    @IsEmail()
    @IsNotEmpty()
    email: string;

    @ApiProperty()
    @IsPhoneNumber()
    @IsNotEmpty()
    phone: string;

    @ApiProperty()
    @IsString()
    @IsNotEmpty()
    username: string;

    @ApiProperty()
    @IsNumber()
    @IsNotEmpty()
    total: number;

    @ApiProperty()
    voucher: any;

    @ApiProperty({
        type: 'array',
        items: {
            type: 'object',
            properties: {
                milkId: { type: 'string' },
                name: { type: 'string' },
                quantity: { type: 'number' },
                price: { type: 'number' }
            }
        }
    })
    @IsString()
    @IsNotEmpty()
    items: OrderItemEntity[];

    payment?: any;

}