# swagger-spec.json
{"openapi":"3.0.0","paths":{"/file/upload":{"post":{"operationId":"ImageController_uploadFile","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"201":{"description":""}},"tags":["file"],"security":[{"bearer":[]}]}},"/file/image/{path}":{"get":{"operationId":"ImageController_showFile","parameters":[{"name":"path","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["file"]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDataDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginLocalDataDto"}}}},"responses":{"200":{"description":""}},"tags":["auth"]}},"/auth/login-firebase":{"post":{"operationId":"AuthController_loginWithFirebaseToken","parameters":[{"name":"authorization","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["auth"]}},"/auth/verify-token":{"post":{"operationId":"AuthController_verifyToken","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"],"security":[{"bearer":[]}]}},"/auth/verify-admin-token":{"post":{"operationId":"AuthController_verifyAdminToken","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"],"security":[{"bearer":[]}]}},"/auth/admin/all-accounts":{"get":{"operationId":"AuthController_getAllAccounts","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"],"security":[{"bearer":[]}]}},"/auth/admin/create-account":{"post":{"operationId":"AuthController_createAccount","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAccountDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"],"security":[{"bearer":[]}]}},"/auth/admin/update-account/{id}":{"put":{"operationId":"AuthController_updateAccount","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminUpdateAccountDataDto"}}}},"responses":{"200":{"description":""}},"tags":["auth"],"security":[{"bearer":[]}]}},"/auth/admin/delete-account/{id}":{"delete":{"operationId":"AuthController_deleteAccount","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["auth"],"security":[{"bearer":[]}]}},"/auth/user/change-password":{"post":{"operationId":"AuthController_changePassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordDto"}}}},"responses":{"200":{"description":""}},"tags":["auth"],"security":[{"bearer":[]}]}},"/auth/user/update-phone":{"post":{"operationId":"AuthController_changePhoneNumber","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhoneDto"}}}},"responses":{"200":{"description":""}},"tags":["auth"],"security":[{"bearer":[]}]}},"/auth/user/update-username":{"post":{"operationId":"AuthController_changeUsername","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsernameDto"}}}},"responses":{"200":{"description":""}},"tags":["auth"],"security":[{"bearer":[]}]}},"/milk":{"get":{"operationId":"MilkController_getAll","parameters":[],"responses":{"200":{"description":""}},"tags":["milk"],"security":[{"bearer":[]}]},"post":{"operationId":"MilkController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMilkDto"}}}},"responses":{"201":{"description":""}},"tags":["milk"],"security":[{"bearer":[]}]}},"/milk/{id}":{"get":{"operationId":"MilkController_getOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["milk"],"security":[{"bearer":[]}]},"put":{"operationId":"MilkController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMilkDto"}}}},"responses":{"200":{"description":""}},"tags":["milk"],"security":[{"bearer":[]}]},"delete":{"operationId":"MilkController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["milk"],"security":[{"bearer":[]}]}},"/milk/undelete/{id}":{"put":{"operationId":"MilkController_unDelete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["milk"],"security":[{"bearer":[]}]}},"/milk-user":{"get":{"operationId":"MilkUserController_getAll","parameters":[],"responses":{"200":{"description":""}},"tags":["milk-user"]}},"/milk-user/{id}":{"get":{"operationId":"MilkUserController_getOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["milk-user"]}},"/voucher":{"get":{"operationId":"VoucherController_getAll","parameters":[],"responses":{"200":{"description":""}},"tags":["voucher"],"security":[{"bearer":[]}]},"post":{"operationId":"VoucherController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateVoucherDto"}}}},"responses":{"201":{"description":""}},"tags":["voucher"],"security":[{"bearer":[]}]}},"/voucher/{id}":{"get":{"operationId":"VoucherController_getOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["voucher"],"security":[{"bearer":[]}]},"put":{"operationId":"VoucherController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateVoucherDto"}}}},"responses":{"200":{"description":""}},"tags":["voucher"],"security":[{"bearer":[]}]},"delete":{"operationId":"VoucherController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["voucher"],"security":[{"bearer":[]}]}},"/voucher/undelete/{id}":{"put":{"operationId":"VoucherController_unDelete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["voucher"],"security":[{"bearer":[]}]}},"/voucher-user":{"get":{"operationId":"VoucherUserController_getAll","parameters":[],"responses":{"200":{"description":""}},"tags":["voucher-user"]}},"/voucher-user/{id}":{"get":{"operationId":"VoucherUserController_getOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["voucher-user"]}},"/category":{"get":{"operationId":"CategoryController_getAll","parameters":[],"responses":{"200":{"description":""}},"tags":["category"],"security":[{"bearer":[]}]},"post":{"operationId":"CategoryController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","example":"Category name"}}}}}},"responses":{"201":{"description":""}},"tags":["category"],"security":[{"bearer":[]}]}},"/category/{id}":{"get":{"operationId":"CategoryController_getOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["category"],"security":[{"bearer":[]}]},"put":{"operationId":"CategoryController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","example":"Category name"}}}}}},"responses":{"200":{"description":""}},"tags":["category"],"security":[{"bearer":[]}]},"delete":{"operationId":"CategoryController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["category"],"security":[{"bearer":[]}]}},"/category/undelete/{id}":{"put":{"operationId":"CategoryController_unDelete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["category"],"security":[{"bearer":[]}]}},"/category-user":{"get":{"operationId":"CategoryUserController_getAll","parameters":[],"responses":{"200":{"description":""}},"tags":["category-user"]}},"/category-user/{id}":{"get":{"operationId":"CategoryUserController_getOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["category-user"]}},"/order":{"post":{"operationId":"OrderController_createOrder","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderDto"}}}},"responses":{"201":{"description":""}},"tags":["order"]}},"/order/get-result/{orderId}":{"get":{"operationId":"OrderController_getResult","parameters":[{"name":"orderId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["order"]}}},"info":{"title":"Milk - API - Documentation","description":"API for user interface development","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterDataDto":{"type":"object","properties":{"username":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"password":{"type":"string"}},"required":["username","email","phone","password"]},"LoginLocalDataDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"CreateAccountDto":{"type":"object","properties":{"username":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"password":{"type":"string"},"role":{"type":"string"}},"required":["username","email","phone","password","role"]},"AdminUpdateAccountDataDto":{"type":"object","properties":{"username":{"type":"string"},"phone":{"type":"string"},"role":{"type":"string"},"password":{"type":"string"}},"required":["username","phone","role","password"]},"PasswordDto":{"type":"object","properties":{"value":{"type":"string"}},"required":["value"]},"PhoneDto":{"type":"object","properties":{"value":{"type":"string"}},"required":["value"]},"UsernameDto":{"type":"object","properties":{"value":{"type":"string"}},"required":["value"]},"CreateMilkDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"},"quantity":{"type":"number"},"images":{"type":"array","items":{"type":"string"}},"categoryId":{"type":"string"},"expiredAt":{"format":"date-time","type":"string"}},"required":["name","description","price","quantity","images","categoryId","expiredAt"]},"MilkStatus":{"type":"string","description":"Milk status","enum":["ACTIVE","INACTIVE"]},"UpdateMilkDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"},"quantity":{"type":"number"},"images":{"type":"array","items":{"type":"string"}},"status":{"$ref":"#/components/schemas/MilkStatus"}},"required":["name","description","price","quantity","images","status"]},"CreateVoucherDto":{"type":"object","properties":{"code":{"type":"string"},"discount":{"type":"number"},"quantity":{"type":"number"},"startAt":{"format":"date-time","type":"string"},"expiredAt":{"format":"date-time","type":"string"}},"required":["code","discount","quantity","startAt","expiredAt"]},"UpdateVoucherDto":{"type":"object","properties":{"code":{"type":"string"},"discount":{"type":"number"},"quantity":{"type":"number"},"startAt":{"type":"string"},"expiredAt":{"type":"string"},"isPublic":{"type":"boolean"}},"required":["code","discount","quantity","startAt","expiredAt","isPublic"]},"CreateOrderDto":{"type":"object","properties":{"email":{"type":"string"},"phone":{"type":"string"},"username":{"type":"string"},"total":{"type":"number"},"voucher":{"type":"object"},"items":{"type":"array","items":{"type":"object","properties":{"milkId":{"type":"string"},"name":{"type":"string"},"quantity":{"type":"number"},"price":{"type":"number"}}}}},"required":["email","phone","username","total","voucher","items"]}}}}